# sumi-daemon/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sumi-daemon | v1",
    "description": "Utoipa's axum bindings for seamless integration for the two",
    "contact": {
      "name": "Juha Kukkonen",
      "email": "juha7kukkonen@gmail.com"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:27381/"
    }
  ],
  "paths": {
...
    "/api/v1/app/lan": {
      "get": {
        "tags": [
          "app"
        ],
        "summary": "`GET /api/v1/app/lan`：读取 [web] 配置（admin 限定；viewer 拿到 write_token 即提权）",
        "operationId": "lan_get",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_LanConfig"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "app"
        ],
        "summary": "`PUT /api/v1/app/lan`：写回 [web] 配置（校验 + toml_edit 保注释写回；失败不落盘）",
        "operationId": "lan_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LanPut"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_LanConfig"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Envelope_LanConfig": {
        "type": "object",
        "description": "统一成功信封；data 为空时省略该字段",
        "required": [
          "status"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "enabled",
              "port",
              "token",
              "writable",
              "separate_write_token",
              "write_token",
              "active"
            ],
            "properties": {
              "active": {
                "type": "boolean"
              },
              "enabled": {
                "type": "boolean"
              },
              "port": {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              },
              "separate_write_token": {
                "type": "boolean"
              },
              "token": {
                "type": "string"
              },
              "writable": {
                "type": "boolean"
              },
              "write_token": {
                "type": "string"
              }
            }
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "LanPut": {
        "type": "object",
        "required": [
          "enabled",
          "port",
          "token",
          "writable",
          "separate_write_token",
          "write_token"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "port": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "separate_write_token": {
            "type": "boolean"
          },
          "token": {
            "type": "string"
          },
          "writable": {
            "type": "boolean"
          },
          "write_token": {
            "type": "string"
          }
        }
      },
...
    }
  }
}